move pytest options from pytest config to tox (#156)
The --cov option was useless, since tox already uses `coverage run`
instead. While we are at it, we can also drop pytest-cov from the test
dependencies. One less wheel for tox to download and install.

The timeout option is moved to tox.

Both of these options add additional dependencies for running `pytest`
outside of tox, which may sometimes be desired if you really know what
you are doing. There is no way to disable them other than by patching
out the addopts configuration. Best to specify this only in tox, which
is designed to set up the appropriate environments already.

Bug: https://bugs.gentoo.org/922192